+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Deactivated asset profiles automatically on delisting in the _Yahoo Finance_ service
 - Optimized the query of the data range functionality (`getRange()`) in the market data service
+- Moved the subscription offer from the info to the user service
 - Upgraded `Nx` from version `20.7.1` to `20.8.0`
 - Upgraded `prisma` from version `6.5.0` to `6.6.0`
 - Upgraded `storybook` from version `8.4.7` to `8.6.12`